+++ apheleia-formatters.el
@@ -206,13 +206,13 @@
                       (sh "posix")
                       (t "bash"))
               (when apheleia-formatters-respect-indent-level
-                (list
-                 "-i" (number-to-string
-                       (cond
-                        (indent-tabs-mode 0)
-                        ((boundp 'sh-basic-offset)
-                         sh-basic-offset)
-                        (t 4)))))
+                (format
+                 "--indent=%d"
+                 (cond
+                  (indent-tabs-mode 0)
+                  ((boundp 'sh-basic-offset)
+                   sh-basic-offset)
+                  (t 4))))
               "-"))
     (rufo . ("rufo" "--filename" filepath "--simple-exit"))
     (stylua . ("stylua" "-"))