+++ test/formatters/samplecode/runic/in.jl
@@ -1,0 +1,7 @@
+
+dist(p1,p2)=sqrt((p1[1]-p2[1])^2+(p1[2]-p2[2]^2))
+
+function main(a,b)
+    return c =
+    magicalfunction(dist(a,b))
+end