+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog].
 ## Unreleased
 ### Formatters
 * Use clang-format for formatting Objective-C/C++ files ([#378]).
+* `bibtex-reformat` for BibTeX files.
 
 ## 4.4.2 (released 2025-11-21)
 ### Bugs fixed